test: :test_tube: migrate tests for AllowedERC725YKeys
