+++ tests/LSP6KeyManager/effects/index.ts
@@ -8,3 +8,5 @@ export * from "./PermissionTransferValue.test";
 export * from "./PermissionSign.test";
 export * from "./AllowedAddresses.test";
 export * from "./AllowedFunctions.test";
+export * from "./AllowedStandards.test";
+export * from "./AllowedERC725YKeys.test";
